Create README.md
With the help of Java and Android Studio, I have created this app which helps to visualize Kruskal's Algorithm. This application helps to understand the working behind Kruskal's algorithm in easy steps to find the minimum spanning tree. I have divided the visualization mainly into two parts: The first part helps to visualize how we remove the edges from the graph in ascending order according to edge weights assigned. And the second step helps to visualize how we put back the edges without forming a cycle in between the nodes. If any cycle is formed then that edge is skipped.